Ensure clangd is executed in vscode.workspace.rootPath (#153)
If the rootPath of a project is a symlink vscode can execute clangd in
the users home-directory. This can cause problems if one is, for
example, using a wrapper script which relies on knowing the current
project. This solves this issue by forcing the clangd to execute in
`vscode.workspace.rootPath` if available.

Test: manual