Inlay hints: support position & arbitrary kinds. (#276)
Previously we had hardcoded rules for which inlay hint kinds were
prefixes vs suffixes. Explicit location information was added in
https://github.com/llvm/llvm-project/commit/7c19fdd59939249c23384f0900d49aab4a5f0695
so we use this.

This enables us to render unknown hint kinds, too.
Instead of a struct { parameterHints, typeHints } we now have a map, and
create decoration types for them on demand.
So far these all render the same, but we should likely give them
different colors.

The TextEditorDecorationTypes were previously coupled to whether decorations
were prefixes or suffixes. However this isn't necessary: the
"attachments" are not rendered if empty, so it's enough to set the text
on the decoration itself.