Disable commit-characters for all code completion.
The use of commit-characters in clangd had good intentions, but it has
terrible UX in practice in VSCode, as documented on the linked bug.

We didn't notice this until it was far too late: the commit characters
were added in clangd 12, but bugs in vscode-languageclient prevented
them from being noticed until languageclient-8.
(https://github.com/microsoft/vscode-languageserver-node/issues/673)

At this point we can't go back and change the server behavior so we need
to override it on the client.

Fixes https://github.com/clangd/vscode-clangd/issues/357