Enable clangd on HLSL sources
Clang is gaining support for HLSL. This change enables using clangd for
the language server for vscode.

I'm unsure if there is a good way to gate this support on a specific
version of clang, but when used with clang-15 it produces an error at
the beginning of the file noting that compilation failed. It doesn't do
anything too terrible if clang can't handle HLSL, so it might be safe to
enable everywhere as-is.