Add support for multi-root workspaces
These changes add proper support for multi root workspaces with a
clangd server per workspace folder.

On a high level, we have a single language client per folder, and
have to rework things so that various global objects (views, ...)
are actually global and not instantiated per client anymore.

The important use case this enables is where clangd is invoked within
a container per process with all the dependencies installed in the
container. To make this use case work in a multi-root setup, a clangd
instance per workspace folder has to be spawned.

Fixes #38